Encapsulate selection functionality in LabelWithSelectionWidget
