+++ OpenRA.Mods.Common/Widgets/LabelWidget.cs
@@ -27,7 +27,6 @@ namespace OpenRA.Mods.Common.Widgets
 		public TextVAlign VAlign = TextVAlign.Middle;
 		public string Font = ChromeMetrics.Get<string>("TextFont");
 		public Color TextColor = ChromeMetrics.Get<Color>("TextColor");
-		public Color TextColorHighlight = ChromeMetrics.Get<Color>("TextfieldColorHighlight");
 		public bool Contrast = ChromeMetrics.Get<bool>("TextContrast");
 		public bool Shadow = ChromeMetrics.Get<bool>("TextShadow");
 		public Color ContrastColorDark = ChromeMetrics.Get<Color>("TextContrastColorDark");
@@ -36,17 +35,13 @@ namespace OpenRA.Mods.Common.Widgets
 		public bool WordWrap = false;
 		public Func<string> GetText;
 		public Func<Color> GetColor;
-		public Func<Color> GetColorHighlight;
 		public Func<Color> GetContrastColorDark;
 		public Func<Color> GetContrastColorLight;
-
-		static Selection selection = new Selection();
 
 		public LabelWidget()
 		{
 			GetText = () => Text;
 			GetColor = () => TextColor;
-			GetColorHighlight = () => TextColorHighlight;
 			GetContrastColorDark = () => ContrastColorDark;
 			GetContrastColorLight = () => ContrastColorLight;
 		}
@@ -67,7 +62,6 @@ namespace OpenRA.Mods.Common.Widgets
 			WordWrap = other.WordWrap;
 			GetText = other.GetText;
 			GetColor = other.GetColor;
-			GetColorHighlight = other.GetColorHighlight;
 			GetContrastColorDark = other.GetContrastColorDark;
 			GetContrastColorLight = other.GetContrastColorLight;
 		}
@@ -148,56 +142,5 @@ namespace OpenRA.Mods.Common.Widgets
 		}
 
 		public override Widget Clone() { return new LabelWidget(this); }
-
-		public override bool HandleMouseInput(MouseInput mi)
-		{
-			switch (mi.Event)
-			{
-				case MouseInputEvent.Down:
-					return selection.HandleMouseDown(this, mi.Location);
-				case MouseInputEvent.Move:
-					return selection.HandleMouseMove(mi.Location);
-				case MouseInputEvent.Up:
-					// Can't copy stuff if we don't have keyboard focus!
-					if (!RenderBounds.Contains(mi.Location) || !TakeKeyboardFocus())
-						return false;
-
-					return selection.HandleMouseUp();
-				default:
-					throw new Exception("Unrecognized MouseEvent on Label.");
-			}
-		}
-
-		public override void MouseExited()
-		{
-			selection.HandleMouseExit();
-		}
-
-		public override bool HandleKeyPress(KeyInput e)
-		{
-			if (e.Event == KeyInputEvent.Up)
-				return false;
-
-			var isOSX = Platform.CurrentPlatform == PlatformType.OSX;
-
-			if ((!isOSX && !e.Modifiers.HasModifier(Modifiers.Ctrl)) || (isOSX && !e.Modifiers.HasModifier(Modifiers.Meta)))
-				return false;
-
-			if (e.Key != Keycode.C)
-				return false;
-
-			if (!selection.OwnedBy(this))
-				return false;
-
-			if (selection.State == Selection.States.Empty)
-				return false;
-
-			if (string.IsNullOrEmpty(selection.SelectedText))
-				return false;
-
-			Game.Renderer.SetClipboardText(selection.SelectedText);
-
-			return true;
-		}
 	}
 }
