+++ mods/common/chrome/lobby.yaml
@@ -6,7 +6,7 @@ Background@SERVER_LOBBY:
 	Height: 600
 	Children:
 		ColorPreviewManager@COLOR_MANAGER:
-		Label@SERVER_NAME:
+		LabelWithSelection@SERVER_NAME:
 			Y: 16
 			Align: Center
 			Width: PARENT_RIGHT
@@ -108,17 +108,17 @@ Background@SERVER_LOBBY:
 							Width: PARENT_RIGHT - 27
 							Height: 16
 							Children:
-								Label@TIME:
+								LabelWithSelection@TIME:
 									X: 3
 									Width: 50
 									Height: 16
 									Shadow: True
-								Label@NAME:
+								LabelWithSelection@NAME:
 									X: 45
 									Width: 50
 									Height: 16
 									Shadow: True
-								Label@TEXT:
+								LabelWithSelection@TEXT:
 									X: 55
 									Width: PARENT_RIGHT - 60
 									Height: 16
